Continue socket recv in same thread without epoll overhead for write RPCs
Current scenario:
Xprt socket fd is being monitored via epoll mechanism for the arrival of data on the socket.
Once the data is arrived, an epoll event is generated and the socket fd is removed from the epoll fd.
A new thread is chosen to monitor the epoll fd while the current thread recv 1 RPC worth of data.
In the current thread, once the data is received, the socket fd is added back to the epoll fd
so that other thread can start recving data while the current thread than continues executing the RPC.

Changed scenario:
We dont add the socket fd back to epoll fd unconditionally once the data is read from the socket.
We look at the RPC size, if it's more than 4KB (which can only happen in the case of write RPC)
and if there is data still available on the socket fd to be read, the current thread
will keep doing recv while dispatching the requests asynchronously to svc_request_async().
At max we will recv max_cont_recv requests continuously (defaults to 64) in a single thread.
With this patch we are seeing consistent seq write perf improvement in the range of 15 - 20%